requirements.txt
absl-py==2.4.0
certifi==2026.4.22
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.9
mediapipe==0.10.35
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.2
pillow==12.2.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-uinput==1.0.1
six==1.17.0
sounddevice==0.5.5
